Refactor diagram
