+++ src/screens/simulation/blocks/index.js
@@ -1,0 +1,5 @@
+import TotalPopulationBlock from './TotalPopulation';
+import ExposedPopulationBlock from './ExposedPopulation';
+import AverageDurationBlock from './AverageDuration';
+
+export { TotalPopulationBlock, ExposedPopulationBlock, AverageDurationBlock };
