+++ src/screens/simulation/colors.js
@@ -1,0 +1,26 @@
+export default {
+    incubation: {
+        main: 'rgb(164, 18, 179)',
+        bg: 'rgba(164, 18, 179, 0.6)',
+    },
+    normal_care: {
+        main: 'rgb(255, 88, 132)',
+        bg: 'rgba(255, 88, 132, 0.6)',
+    },
+    intensive_care: {
+        main: 'rgb(54, 162, 235)',
+        bg: 'rgba(54, 162, 235, 0.6)',
+    },
+    following_care: {
+        main: 'rgb(54, 54, 255)',
+        bg: 'rgba(54, 54, 255, 0.6)',
+    },
+    death: {
+        main: 'rgb(88, 88, 88)',
+        bg: 'rgba(88, 88, 88, 0.6)',
+    },
+    recovered: {
+        main: 'rgb(88, 235, 88)',
+        bg: 'rgba(88, 235, 88, 0.6)',
+    },
+};
